+++ packages/backend-rs/src/model/entity/drive_file.rs
@@ -48,10 +48,6 @@ pub struct Model {
     pub request_headers: Option<Json>,
     #[sea_orm(column_name = "requestIp")]
     pub request_ip: Option<String>,
-    #[sea_orm(column_name = "maybeSensitive")]
-    pub maybe_sensitive: bool,
-    #[sea_orm(column_name = "maybePorn")]
-    pub maybe_porn: bool,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]