+++ packages/backend/src/models/entities/user-profile.ts
@@ -186,11 +186,6 @@ export class UserProfile {
 	@Column("boolean", {
 		default: false,
 	})
-	public autoSensitive: boolean;
-
-	@Column("boolean", {
-		default: false,
-	})
 	public carefulBot: boolean;
 
 	@Column("boolean", {